renamed geojson.py to geojson_wrapper.py and updated throughout
