+++ strabo/strabo/views/private.py
@@ -11,7 +11,7 @@ edit_ip, edit_event, get_geojson, insert_text, edit_textselection
 from strabo.utils import make_date, DMS_to_Dec, clean_date, prettify_columns, \
 get_raw_column, get_fields
 from strabo.image_processing import make_thumbnail, allowed_file #, getEXIF
-from strabo.geojson import get_coords, get_type, add_name_and_color, \
+from strabo.geojson_wrapper import get_coords, get_type, add_name_and_color, \
 make_featureCollection
 from strabo.filewriting import write_to, rewrite_geojson
